feat: Es links use FieldConfig Version2 schema；use RenderValue to display
